requirements.txt
aiohttp>=3.9.0
google-auth>=2.27.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.120.0
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
aioresponses>=0.7.6
